split init code into services subpackage
